refactor: reorganize overview components into v1 and v2 folders
- Move shared Overview component to v1/features/overview/ (no longer shared)
- Move v2 overview content from workspaces/overview/ to overview/ folder
- Update imports to reflect new component locations
- Remove duplicate WorkspacesOverview component

The shared overview component was only used by v1, so it's been moved
to the v1 folder. The v2 overview is now properly in the Overview island
instead of the Workspaces island.
