+++ src/v1/features/overview/overview.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Overview } from '../../../shared/components/overview/Overview';
+import { Overview } from './OverviewContent';
 import pathnames from '../../utilities/pathnames';
 
 const V1Overview: React.FC = () => <Overview links={{ groups: pathnames.groups.link(), roles: pathnames.roles.link() }} />;
